feat: Add suffle_answers field to Question model
The Question model has been updated to include a new field called suffle_answers. This field is a boolean that determines whether the answers for a question should be shuffled or not. By default, suffle_answers is set to True, meaning that the answers will be shuffled. This change enhances the functionality of the application by providing the ability to control the order of the answers for each question.