feat: Add suffle_answers field to QuestionSerializer
The `QuestionSerializer` in `serializers.py` has been updated to include a new field called `suffle_answers`. This field is added to the `fields` list and allows for the shuffling of answers for each question. This change enhances the functionality of the application by providing more control over the order of the answers.